serde settings
