+++ crates/air/Cargo.toml
@@ -13,6 +13,7 @@ luminal.workspace = true
 tracing.workspace = true
 numerair.workspace = true
 serde.workspace = true
+serde_json.workspace = true
 thiserror.workspace = true
 num-traits.workspace = true
 typetag.workspace = true
