+++ crates/prelude/src/prelude.rs
@@ -2,16 +2,15 @@
 pub use luminal::prelude::*;
 pub use luminal_nn::*;
 
+// --- luminair_utils ---
+pub use luminair_utils::{JsonDeserialization, JsonSerialization};
+
 // --- luminair_graph ---
 pub use luminair_graph::graph::LuminairGraph;
 pub use luminair_graph::StwoCompiler;
 
 // --- luminair_prover ---
-pub use luminair_prover::{
-    prover::prove,
-    serde::{JsonDeserialization, JsonSerialization},
-    LuminairProof,
-};
+pub use luminair_prover::{prover::prove, LuminairProof};
 
 // --- luminair_verifier ---
 pub use luminair_air::{pie::LuminairPie, settings::CircuitSettings};
